requirements.txt
watchdog>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
click>=8.0.0
PyYAML>=6.0.0
python-dotenv>=1.0.0
pywin32>=306; sys_platform == 'win32'  # For Windows printing
structlog>=24.1.0
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.7.0
